# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dog",
        "description": "Documentação da API Dog",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/posts/{post}/like": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Curtir um post",
                "description": "Curtir um post (resposta rápida via Redis, persistida assincronamente)",
                "operationId": "466e3ae39106d7c67096622459c5e09b",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID ou UUID do post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contagem atual de likes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "likes": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/unlike": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Remover like de um post",
                "description": "Remover like de um post (resposta rápida via Redis, persistida assincronamente)",
                "operationId": "7394a22e92c3def9ca2bfe54d4db36fd",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID ou UUID do post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contagem atual de likes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "likes": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}